stuff
Added a notebook that can display and save a mosaic (you can choose the number of images in the mosaic) of random images with displayed bounding boxes of either a single volunteer or all volunteers.

Also added a function to the model that saves the metrics used to calculate the curves  at each epoch, and the final val run. saved as plots.csv